atualização 2.0v

+++ tipos_variaveis/src/TiposVariaveis.java
@@ -3,7 +3,7 @@ public class TiposVariaveis {
         
         short numeroCurto = 1;
         int numeroNormal = numeroCurto;
-        short numeroLongo = (short) numeroNormal;
+        long numeroLongo = (short) numeroNormal;
 
         int numero = 5;
